app/requirements.txt
faiss-gpu==1.7.2
fastapi==0.85.1
laserembeddings==1.1.2
numpy==1.21.6
pydantic==1.10.2
torch==1.12.1
uvicorn==0.19.0
jieba==0.42.1